Revert "refactor: remove rounding for timestamp"
This reverts commit 856f35f846db11bc3e3b6d2f6f8c0c2846093099.

+++ py_vsys/model.py
@@ -470,7 +470,7 @@ class VSYSTimestamp(NonNegativeInt):
         Returns:
             VSYSTimestamp: The VSYSTimestamp.
         """
-        return cls(time.time() * cls.SCALE)
+        return cls(int(time.time() * cls.SCALE))
 
     @property
     def unix_ts(self) -> float:
